czi-70: more bug fixes
